+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 * Add [example](examples/json) to translate JSON inputs. 
+* Added platform and python version information to the user-agent string that is sent with API calls, along with an opt-out.
+* Added method for applications that use this library to identify themselves in API requests they make.
 
 
 ## [1.13.0] - 2023-01-26
@@ -19,7 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   Note: older library versions also support the new languages, this update only adds new code constants.
 ### Changed
-* Added system, python and request library version information to the user-agent string that is sent with API calls.
 ### Deprecated
 ### Removed
 ### Fixed